+++ distributedai/build.gradle
@@ -68,10 +68,9 @@ dependencies {
 
     implementation 'com.mattskala:itemadapter:0.3'
 
+    implementation 'org.openmined.kotlinsyft:syft:0.1.3'
 
     implementation "com.androidplot:androidplot-core:1.5.7"
-
-    implementation 'org.openmined.kotlinsyft:syft:0.1.3'
 
     // Testing
     testImplementation 'junit:junit:4.12'
