+++ federated_ml/build.gradle
@@ -1,28 +1,120 @@
-plugins {
-    id 'application'
-    id 'kotlin'
-}
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
+apply plugin: 'org.jlleitschuh.gradle.ktlint'
+apply plugin: 'androidx.navigation.safeargs.kotlin'
+apply plugin: 'kotlinx-serialization'
 
+android {
+    compileSdkVersion 29
+    buildToolsVersion "29.0.3"
+
+    useLibrary 'android.test.base'
+    useLibrary 'android.test.mock'
+
+    viewBinding {
+        enabled = true
+    }
+
+    defaultConfig {
+        minSdkVersion 22
+        targetSdkVersion 29
+        versionCode 1
+        versionName "1.0"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles 'consumer-rules.pro'
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+
+// To inline the bytecode built with JVM target 1.8 into
+// bytecode that is being built with JVM target 1.6. (e.g. navArgs)
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = "1.8"
+        allWarningsAsErrors = true
+    }
+
+    testOptions {
+        unitTests.returnDefaultValues = true
+    }
+
+    packagingOptions {
+        exclude 'META-INF/DEPENDENCIES'
+        exclude 'META-INF/LICENSE'
+        exclude 'META-INF/LICENSE.txt'
+        exclude 'META-INF/license.txt'
+        exclude 'META-INF/NOTICE'
+        exclude 'META-INF/NOTICE.txt'
+        exclude 'META-INF/notice.txt'
+        exclude 'META-INF/ASL2.0'
+        exclude("META-INF/*.kotlin_module")
+    }
+}
 
 dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'io.github.microutils:kotlin-logging:1.7.7'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.7'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.7'
-    implementation "com.squareup.sqldelight:sqlite-driver:$sqldelight_version"
-    implementation 'com.mpatric:mp3agic:0.9.1'
-    implementation "com.squareup.sqldelight:coroutines-extensions:$sqldelight_version"
+    implementation project(':common')
+    implementation project(':ipv8-android')
+
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation "androidx.preference:preference:1.1.1"
+    implementation 'androidx.core:core-ktx:1.3.1'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.android.material:material:1.2.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0'
+    testImplementation 'junit:junit:4.12'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.3.0'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.3.0'
+    implementation 'com.google.android:flexbox:2.0.1'
+    implementation 'org.jetbrains.kotlinx:kotlinx-serialization-runtime:0.14.0'
     implementation "com.github.haifengl:smile-kotlin:2.6.0"
 
-    implementation 'org.openmined.kotlinsyft:syft:0.1.3'
-    // https://mvnrepository.com/artifact/org.openmined.kotlinsyft/syft
-    implementation group: 'org.openmined.kotlinsyft', name: 'syft', version: '0.1.3'
+    // BitTorrent/Libtorrent libraries
+    implementation 'com.turn:ttorrent-core:1.5'
 
-    implementation 'org.nield:kotlin-statistics:1.2.1'
-}
+    implementation 'com.google.android.exoplayer:exoplayer-core:2.10.5'
+    implementation 'com.google.android.exoplayer:exoplayer-dash:2.10.5'
+    implementation 'com.google.android.exoplayer:exoplayer-ui:2.10.5'
+    implementation 'com.google.android.exoplayer:exoplayer-hls:2.10.5'
 
-application {
-    mainClassName = 'com.example.federated_ml.TestKt'
+    // Cryptocurrency integration
+    implementation 'org.bitcoinj:bitcoinj-core:0.15.7'
+//    implementation 'com.google.guava:guava:29.0-android'
+    implementation 'org.knowm.xchange:xchange-parent:5.0.1'
+    implementation 'org.knowm.xchange:xchange-binance:5.0.1'
+
+    // Crypto library for testing
+    testImplementation "com.goterl.lazycode:lazysodium-java:4.2.4"
+
+    // Testing
+    testImplementation 'junit:junit:4.12'
+    testImplementation "io.mockk:mockk:1.9.3"
+    testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.3'
+    testImplementation "org.robolectric:robolectric:3.4.2"
+
+    // Testing and generating example data
+    implementation "com.squareup.sqldelight:sqlite-driver:$sqldelight_version"
+
+    // Reading MP3 metadata
+    implementation 'com.mpatric:mp3agic:0.9.0'
+
+    // Logging
+    implementation 'io.github.microutils:kotlin-logging:1.7.7'
+
+    implementation project (':federated_ml')
 }
 
 tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
@@ -31,3 +123,12 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
         "-Werror" // Set Kotlin compiler warnings as errors
     ]
 }
+
+tasks.withType(Test) {
+    systemProperty "java.library.path", "./src/main/jniLibs"
+}
+
+application {
+    mainClassName = 'com.example.federated_ml.TestKt'
+}
+
